Eject oauth-store from the core cycle
The OAuth connect/disconnect paths reached into home/suggested-prompts.ts
to invalidate the Home-feed suggestion cache, and suggested-prompts.ts
loops back to oauth-store transitively through its heavy generation deps
(system-prompt builder -> prompts/sections -> ... -> oauth-store). That
mutual reachability trapped oauth-store plus ~22 downstream modules in the
daemon-core SCC.

Split the checkpoint-backed cache (read/write/invalidate) out of
suggested-prompts.ts into a dependency-light leaf, home/suggested-prompts-cache.ts,
that depends only on the checkpoint store and the event hub — neither of
which reaches oauth-store. Repoint oauth-store and connect-orchestrator at
the leaf; oauth-store's invalidation can now be a plain static import
instead of the cycle-dodging dynamic import (the stale comment claiming a
direct back-import is dropped). suggested-prompts.ts imports its own cache
helpers from the leaf.

Core SCC: 287 -> 264 files (-23). oauth-store, suggested-prompts, and the
new cache leaf are all out of every cycle. No behavioral change.
