+++ assistant/src/home/suggested-prompts.ts
@@ -17,63 +17,20 @@
 
 import { resolveCallSiteConfig } from "../config/llm-resolver.js";
 import { getConfig } from "../config/loader.js";
-import {
-  deleteMemoryCheckpoint,
-  getMemoryCheckpoint,
-  setMemoryCheckpoint,
-} from "../persistence/checkpoints.js";
 import { buildSystemPrompt } from "../prompts/system-prompt.js";
 import { getConfiguredProvider } from "../providers/provider-send-message.js";
-import { buildAssistantEvent } from "../runtime/assistant-event.js";
-import { assistantEventHub } from "../runtime/assistant-event-hub.js";
 import { runBtwSidechain } from "../runtime/btw-sidechain.js";
 import { formatIntegrationSummary } from "../schedule/integration-status.js";
 import { getLogger } from "../util/logger.js";
 import type { SuggestedPrompt } from "./feed-types.js";
+import {
+  readCachedPrompts,
+  writeCachedPrompts,
+} from "./suggested-prompts-cache.js";
 
 const log = getLogger("suggested-prompts");
 
 const LLM_SUGGESTIONS_TIMEOUT_MS = 5_000;
-const LLM_CACHE_TTL_MS = 4 * 60 * 60 * 1000; // 4 hours
-
-const CHECKPOINT_KEY_JSON = "home:suggested_prompts:json";
-const CHECKPOINT_KEY_TIMESTAMP = "home:suggested_prompts:cached_at";
-
-// ---------------------------------------------------------------------------
-// Checkpoint-backed cache for LLM-generated suggestions
-// ---------------------------------------------------------------------------
-
-function readCachedPrompts(): SuggestedPrompt[] | null {
-  try {
-    const json = getMemoryCheckpoint(CHECKPOINT_KEY_JSON);
-    const timestampStr = getMemoryCheckpoint(CHECKPOINT_KEY_TIMESTAMP);
-    if (!json || !timestampStr) {
-      return null;
-    }
-    const cachedAt = Number(timestampStr);
-    if (isNaN(cachedAt) || Date.now() - cachedAt > LLM_CACHE_TTL_MS) {
-      return null;
-    }
-    const parsed = JSON.parse(json);
-    if (!Array.isArray(parsed)) {
-      return null;
-    }
-    return parsed as SuggestedPrompt[];
-  } catch {
-    return null;
-  }
-}
-
-function writeCachedPrompts(prompts: SuggestedPrompt[]): boolean {
-  try {
-    setMemoryCheckpoint(CHECKPOINT_KEY_JSON, JSON.stringify(prompts));
-    setMemoryCheckpoint(CHECKPOINT_KEY_TIMESTAMP, String(Date.now()));
-    return true;
-  } catch {
-    // Cache write failure is non-fatal — the next refresh regenerates.
-    return false;
-  }
-}
 
 /**
  * Return cached assistant-generated prompts. No LLM calls happen in
@@ -82,36 +39,6 @@ function writeCachedPrompts(prompts: SuggestedPrompt[]): boolean {
  */
 export async function getSuggestedPrompts(): Promise<SuggestedPrompt[]> {
   return readCachedPrompts() ?? [];
-}
-
-/**
- * Drops the suggestion cache so the next on-demand refresh regenerates
- * prompts with current integration state.
- *
- * Called from OAuth connect/disconnect paths so suggestions reflect the
- * new state within one reload instead of waiting for the TTL.
- */
-export function invalidateAssistantSuggestedPromptsCache(): void {
-  try {
-    deleteMemoryCheckpoint(CHECKPOINT_KEY_JSON);
-    deleteMemoryCheckpoint(CHECKPOINT_KEY_TIMESTAMP);
-  } catch {
-    // Invalidation failure is non-fatal — the TTL still bounds staleness.
-  }
-  assistantEventHub
-    .publish(
-      buildAssistantEvent({
-        type: "home_feed_updated",
-        updatedAt: new Date().toISOString(),
-        newItemCount: 0,
-      }),
-    )
-    .catch((err) => {
-      log.warn(
-        { err },
-        "Failed to publish home_feed_updated after prompt cache invalidation",
-      );
-    });
 }
 
 /**
